^r Extract TOML policy load/merge from render_injection_bundle.go, move-only (gofmt, vet, build, package unit+fuzz tests pass; declaration set verified identical; internal restructure, no behavior change)
Step 1 of 3 for the injection god-file split: policy loading, merging,
include validation, and TOML-subset parsing move verbatim to
render_policy_load.go.
